jax and optax as optional
